blissful endianness-opacity
not caring about endianness is the opiate of the masses.
happy, happy masses. remove endianness.h and all its baleful
influence by explicitly breaking up the cell structure. #892

+++ include/ncpp/NotCurses.hh
@@ -238,7 +238,7 @@ namespace ncpp
 			return notcurses_getc_nblock (nc, ni);
 		}
 
-		char* get_at (int yoff, int xoff, uint32_t* attr, uint64_t* channels) const noexcept
+		char* get_at (int yoff, int xoff, uint16_t* attr, uint64_t* channels) const noexcept
 		{
 			return notcurses_at_yx (nc, yoff, xoff, attr, channels);
 		}
